Revert tough_scrolling_cases logic change for Linux platform
Commit aabd0e20f4b3b335dc556f37dbb136538e10c7c7 introduced some
 performance regressions in Linux.

Reverting the changes made to tough_scrollbar_scrolling_cases for platforms that aren't Windows, due to a performance regression issue on Linux. The changes were implemented to account for Fluent scrollbars which aren't available in any platform other than Windows.

The root cause for the regression isn't clear at the moment and a bug was filed to investigate further and unify the two code paths (1445774).
More details in the linked bugs.

Bug: 1444433, 1444767, 1445774
Change-Id: I872456d5e0ae9191f98d22907c5fd9d22efcb20b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4521686
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Commit-Queue: Gaston Rodriguez <gastonr@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1144779}