[Extensions] Allow Webstore to receive extension events
This CL changes the check we have when sending out extension events to
a target context from completely disallowing them to be sent to web page
context, to allow an exception for the Webstore domain. Previously this
wasn't an issue on the old Webstore, as it technically is seen as the
context of a hosted app, but with the new Webstore it is actually
treated as a web page which the given APIs are exposed on.

Also adds a test to ensure a listener can be registered and received for
one of these events.

Bug: 1441136
Change-Id: I013f0706bb5c0a15f487f13072d7206d79f7053a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4523079
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Commit-Queue: Tim <tjudkins@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1144809}