[coverage] Add convenience script for downloading fuzzing corpora.
This patch is part of a series, and the overall goal of the series is to
provide a set of convenience scripts to make it simpler to generate
fuzzing coverage reports.

A subsequent CL will run these scripts in automation but is outside the
scop eof this patch series.

Change-Id: I16f9c080a12f4fdda761c648ede39befdb4e8891
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4529217
Reviewed-by: Prakhar Asthana <pasthana@google.com>
Commit-Queue: Julia Hansbrough <flowerhack@google.com>
Cr-Commit-Position: refs/heads/main@{#1144856}