capture-game: Starts the client session on session init completed
This CL adds callback in `CaptureModeController::Start()` to be
invoked anytime when the state of the session initialization process
is determined. Client session can only start if it's informed to
proceed.
This is necessary since the client session initialization and the
capture session initialization are asynchronous, and the client
session needs to be informed whether to proceed or not when the
session initialization is completed.

Fixed: b/281763025
Change-Id: Ibac5e491a172ed98363a8b375475b6db4ebb3fce
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4525641
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Commit-Queue: Michele Fan <michelefan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1144874}