Allow for redirects to change the network context of a prefetch.
This CL adds support for redirect that require a different network
context than the previous URL of the prefetch. For these types of
redirect, the previous PrefetchStreamingURLLoader is stopped, and a new
PrefetchStreamingURLLoader is used to request the redirect URL in the
appropriate network context.

Bug: 1266876
Change-Id: Ic5d836e193ee94c595cd7e0c7d67bb20f9084232
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4518966
Reviewed-by: Simon Pelchat <spelchat@chromium.org>
Commit-Queue: Max Curran <curranmax@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1144891}