Introducing a new app usage telemetry periodic collector
This change introduces a new periodic collector for app usage telemetry
collection and reporting. This will only control the rate of collection
based on the new policy setting. This is to avoid data staleness because
we do not track timestamps with the telemetry data that is being
persisted in the pref store today. The observer component will perform
necessary validation on allowed types before it persists data in the
pref store.

I will be renaming the AppUsageCollector and all its references
following this change for better distinction in component
responsibilities.

Bug: b:260803048
Change-Id: Idcfd0c428c47086d45527026a8d49743351f9f9d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4535336
Reviewed-by: Leonid Baraz <lbaraz@chromium.org>
Commit-Queue: Vignesh Shenvi <vshenvi@google.com>
Reviewed-by: Ahmed Nasr <anasr@google.com>
Cr-Commit-Position: refs/heads/main@{#1144969}