[Cast Moderator] Refactor AccessCodeCastSinkServiceTest
This CL refactors the AccesscodeCastSinkServiceTest to:
(1) Use MockAccessCodeCastPrefUpdater. In the future, this class will
have two implementations and it's easier to write tests to use a mock
class here.
(2) Clean up the code for testing async methods by using the
TaskEnvironment to control clock and wait for async tasks to complete.

Bug: 1433431
Change-Id: Ibaa2ba31caef0856447f5ee62059d504acab15ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4531288
Commit-Queue: Muyao Xu <muyaoxu@google.com>
Reviewed-by: george benz <gbj@google.com>
Cr-Commit-Position: refs/heads/main@{#1144993}