[no-vary-search-hint] Enable No-Vary-Search hint for origin trial
We would like to experiment with No-Vary-Search hint as part of the
existing origin trial for No-Vary-Search.

We would also like to use "eagerness" functionality to support
No-Vary-Search hint matching between URLs the browser predicts
are good candidates for the next navigation (at pointerdown or
on-hover) and prefetch candidates.

Bug: 1416659
Change-Id: I23d8c8effcf53d966cdb14e07ce004cdcdc28ce2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4520273
Commit-Queue: Liviu Tinta <liviutinta@chromium.org>
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1145058}