[anchor-position] Return the last acceptable anchor element
As decided in https://github.com/w3c/csswg-drafts/issues/8165, when
resolving anchor name conflicts, we should return the last acceptable
anchor element instead of the first one. This patch implements the
change.

The biggest change is that NGLogicalAnchorQuery references are now
a list sorted in reversed tree order, so that the query algorithm
remains the same as before: find the first entry that comes earlier
in tree order.

In in-order vs. out-of-order flag in NGLogicalAnchorQuery creation has
become not very useful because an in-order call simply prepends an
entry to the reference list. Hence this patch removes the flag and the
related logic.

Bug: 1442752
Change-Id: I3f28f1b66097fa7a0a333080dd0f29e62f417b71
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4518226
Reviewed-by: Koji Ishii <kojii@chromium.org>
Commit-Queue: Xiaocheng Hu <xiaochengh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1145068}