Fix DeviceAuthenticator destruction memory issue in auth flow
Currently, we get an instance of DeviceAuthenticator inside of the
function that will use it, but don't keep a reference to
DeviceAuthenticator. This causes the DeviceAuthenticator to be destroyed
immediately, since the factory returns a scoped_refptr of the
DeviceAuthenticator.

Instead, we need any class that uses the DeviceAuthenticator to create a
reference to it and then destroy it after the authentication callback is
complete.

Bug: 4405587, 1445789
Change-Id: I388d9d59a199aeb8556a5a11ba4886e3c19858e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4535337
Reviewed-by: Siyu An <siyua@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Vinny Persky <vinnypersky@google.com>
Reviewed-by: Sujie Zhu <sujiezhu@google.com>
Cr-Commit-Position: refs/heads/main@{#1145079}