Add tap caret to toggle menu behaviour to web contents.
Currently, the touch selection quick menu can be toggled by tapping on
the touch selection handles. This CL allows the menu to also be toggled
by directly tapping on a text caret. The idea is to make the menu easier
to access when no touch handles are visible, e.g. because the user has
typed something.

We implement this in a way similar to how the menu is currently shown
after tapping the selection while touch handles are not visible, i.e.

1. Determine if the caret has been tapped in blink::SelectionController
   and send context menu if needed.
2. Handle the context menu request from
   content::TouchSelectionControllerClientAura by toggling the quick
   menu.

UX visualisation of initially tapping caret to make menu and touch
handles appear:
https://drive.google.com/file/d/1zw35URMgW-7CpA2xzziLZ-I92fOrBJW8/view?usp=sharing

(In this implementation, the toggle menu behaviour also applies to
taps that would keep the caret on the same spot, e.g. for an empty
textfield, this might include tapping the empty area.)

Bug: b:244116654, b:264824232
Change-Id: I600ffaa54ee2c7e641e75d1d53717c4095149e63
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4520088
Reviewed-by: Xiaocheng Hu <xiaochengh@chromium.org>
Commit-Queue: Michelle Chen <michellegc@google.com>
Reviewed-by: Robert Flack <flackr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1145093}