[BRP] Propose renderer-live BRP synthetic trial
This change attempts to be trivial and does not introduce behavioral
change.

The Chrome Memory Team is attempting to enable BRP in the renderer
process. This is the first step in experimentation, in which we set
up the synthetic trial that will house the control and experiment
groups. Both the control and experiment groups seize control of
BRP enablement from the existing experiment.

A future CL will set these flags appropriately to create the groups
for the trial.

Bug: 1444624
Change-Id: I2afdbc235903b747d9293f74fa7e98af9d833a72
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4520187
Commit-Queue: Kalvin Lee <kdlee@chromium.org>
Reviewed-by: Keishi Hattori <keishi@chromium.org>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1145164}