media: Use RAII-like pattern for VADisplayState.
The VADisplayState is a singleton class that keeps track of a VADisplay.
Chrome uses a single VADisplay to make libva calls, and it currently
does manual reference counting to track the initialization state: the
VADisplayState is initialized for the first reference and re-used
thereafter; then it's de-initialized after the last reference is
released. To do this reference counting, users of VADisplayState must
call VADisplayState::Initialize() before using the VADisplayState
singleton and then call VADisplayState::Deinitialize() when they're
done.

Unfortunately, as explained in https://crbug.com/1435500#c41, it's very
easy to forget to call Deinitialize(), and this can actually cause the
GPU process to crash.

This CL refactors the VADisplayState a bit to make it use a RAII-like
pattern. After this CL, users of VADisplayState (now renamed to VADisplayStateSingleton in this CL) no longer need to call
Initialize()/Deinitialize(). Instead, they acquire a
VADisplayStateHandle: if this handle is non-null, users can assume the
VADisplayStateSingleton is initialized. When the VADisplayStateHandle is
destroyed (e.g., by going out of scope), the VADisplayStateSingleton's
reference count is decremented automatically and de-initialization is
performed as needed.

Note: I did consider using a scoped_refptr, but the details make using
this a bit more complicated than it may seem at a first glance: what
we're ref-counting is not the lifetime of an object, but rather its
initialization state. If we use a scoped_refptr, we may hand out a
reference to a VADisplayStateSingleton that will soon be de-initialized.

Note: after this CL, we no longer log vaTerminate() failures for UMA.

Bug: 1435500
Test: on grunt - make sure the GPU process can start after initializing the VADisplayState fails.
Change-Id: I9b3f650114c0125d68062c3c6e3effa280aecb76
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4501962
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Miguel Casas-Sanchez <mcasas@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1145178}