Ash: Extend test coverage for Web Content Handler
This change adds test coverage for the public API
method `RequestLocalApproval` for Ash.

It is a dependency of the change crrev.com/c/4484597,
and it will help refactor the internal methods of
the Web Content Handler that don't need access to
the object's attributes.

Bug: b/278079069
Change-Id: Iec9cd0d7b7905d138113f126858d74b2db7c1f71
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4502790
Commit-Queue: Anthi Orfanou <anthie@google.com>
Reviewed-by: Duong Dac <ddac@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1145200}