[//skia] Set `skia_use_dawn` to true by default (reland)
A prerequisite to running with Graphite is that `skia_use_dawn` be set
to true when building. This CL sets it to true by default in the
production code so that we can ensure that the code is buildable in
this state and facilitate straightforward local running with Graphite.

NOTE: This flag affects only Graphite code paths, so this CL will have
no impact on production code.

This is a reland with fixes for issues exposed on the waterfall in the
diff between PS1 and PS2:
- Dependency issues exposed only in clean builds
- Compile failures exposed only in non-DCHECK builds

Bug: 1423574
Change-Id: I4e7ad03e697821de9722b3ec899f1ff842a8c2d6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4536696
Reviewed-by: Sunny Sachanandani <sunnyps@chromium.org>
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1145231}