media/gpu/chromeos: Handle error if wrapping VideoFrame failed
VideoFrame::WrapVideoFrame() call may fail if the frame has invalid
attributes (size, format, etc.). In that situation, the wrap will fail
and will return an nullptr.

In some cases, the decoding component may send the invalid frame sizes
on purpose, for example in the Android STS. Without the DCHECK enabled,
the Chrome GPU process will crash on the locking attemption. If the GPU
process won't be restarted fast enough, the STS will fail because the
backend won't be able to get any GPU backend.

In any case, crash in such cases is not the best way of handling the
error, so the DCHECK checking for the wrapped frame was replaced with
the if, and the nullptr is returned.


Bug: b:280050397, b:266883226, b:274864105
Test: CtsSecurityTestCases
Test: android.security.cts.StagefrightTest#testStagefright_cve_2019_14132
Change-Id: I68fe90c39d9f52f3305e4134e872b93684b98d41
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4533375
Commit-Queue: Michał Krawczyk <mikrawczyk@google.com>
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Auto-Submit: Michał Krawczyk <mikrawczyk@google.com>
Cr-Commit-Position: refs/heads/main@{#1145244}