[Autofill] Refactor AutofillImageFetcher to support desktop + iOS
Previously, AutofillImageFetcher in //components was a pure interface
with one implementation in //chrome. However, we now plan to add an iOS
implementation as well (in //ios). To reduce code duplication, this CL
refactors AutofillImageFetcher to be an abstract class containing the
'core' fetching logic, and then changes AutofillImageFetcherImpl in
//chrome to only override to provide the specialized logic needed for
desktop.

A future CL will also add an iOS implementation of AutofillImageFetcher,
with appropriate override logic for that platform.

As part of this CL, the network annotation information was
updated with fields that are now required: last_reviewed,
internal::contacts, and user_data::type.

Bug: 1444048
Change-Id: Iadf7c854a19d912a5c87b119d1f60142b47ecb46
Low-Coverage-Reason: Existing test coverage problem, to be fixed in follow-up
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4521407
Reviewed-by: Florian Leimgruber <fleimgruber@google.com>
Auto-Submit: Stephen McGruer <smcgruer@chromium.org>
Commit-Queue: Stephen McGruer <smcgruer@chromium.org>
Reviewed-by: Siyu An <siyua@chromium.org>
Reviewed-by: Martin Šrámek <msramek@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1145335}