Add UMA histogram for causes of irregular LCP image load timing
This CL adds a UMA histogram to count the causes of irregular LCP image
load timing. This is a follow up CL to the previous
https://chromium-review.googlesource.com/c/chromium/src/+/4476487 in
that specific potential causes are being counted. They are
1, image being loaded from memory cache.
2, image being preloaded by the early-hints mechanism.
3, image being loaded from memory cache and preloaded by the early-hints
mechanism.
If none of the above is true when the load timing is incorrect, we
count the cause as unknown.

The information whether an image being loaded from memory cache and/or
preloaded by the early-hints mechanism are propagated from the
ResourceFetcher::RequestResource and passed from renderer to browser.

The UMA recording happens in the PageLoadTracker::OnTimingChanged()
method when a new LCP candidate is about to be updated.

Bug: 1431906
Change-Id: I244a2e63132388d4dede42e54d634ca2fb83c4b9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4514121
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Reviewed-by: Brendon Tiszka <tiszka@chromium.org>
Commit-Queue: Hao Liu <haoliuk@chromium.org>
Reviewed-by: Ian Clelland <iclelland@chromium.org>
Reviewed-by: Joe Mason <joenotcharles@google.com>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1145341}