Introduce backendIdentifier in FormSuggestion
`backendIdentifier` is GUID for credit card and address suggestions, and nil for the other types of popups.

This CL is a prerequisite for https://crrev.com/c/4538542 that tries to remove the reliance on frontend ids from FillorPreviewForm method in the favour of backend ids. Therefore, it is important that the backend ids are correctly set for the suggestions in iOS.

Bug: 1394920
Change-Id: If7e36311c6ca38e0950c8b0910bb4a05b4c12fc0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4538823
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Christoph Schwering <schwering@google.com>
Commit-Queue: Vidhan Jain <vidhanj@google.com>
Cr-Commit-Position: refs/heads/main@{#1145355}