updater: modernize the interface of TaskScheduler for Windows.
This eliminates most most of the runtime CHECKs by replacing pointer
types with reference types for out params.

It adds `nodiscard` for most functions and fixes a couple of call
sites where the return value of the function is being ignored.

The logic of the code was slightly changed in places but not too much.
The intention of the CL is to keep the existing behavior of the
existing code.

Fixed: 1445583
Change-Id: I63d4b3371e34d281189fb36703380d96ea8287ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4540762
Auto-Submit: Sorin Jianu <sorin@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Sorin Jianu <sorin@chromium.org>
Reviewed-by: Joshua Pawlicki <waffles@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1145387}