[Android] Move lists of CachedFlags from ChromeCachedFlags
ChromeFeatureList now contains the lists of CachedFlags (and the
CachedFlags themselves).

It is necessary to have a list of all CachedFlags in
ChromeFeatureList, since we need to map the feature name to the
legacy shared preference key in the CachedFlag when:
1. Setting up test values via @EnableFeatures/@DisableFeatures
    * Currently done through CachedFeatureFlags#setForTesting, but
      requires CachedFeatureFlags to know legacy SharedPrefs key
      overrides, which blocks its migration to //base.
2. Called via Java_CachedFeatureFlags_IsEnabled()
    * Currently uses the default from native passed as a param through
      JNI.

This also adds a test to ensure flags aren't declared and not cached,
which is a common pitfall.

Bug: 1442347
Change-Id: I24092289f438fcc367d9c51a8cb1d7d85e54c5a4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4540276
Commit-Queue: Henrique Nakashima <hnakashima@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1145394}