Use a more plausible RSA size in subtle-crypto-concurrent
The test currently uses 256-bit RSA. For reference, that is so small
that it isn't even in the RSA factoring challenges. 512-bit RSA was
already factored in 1999. Moreover, the test pairs 256-bit RSA with
SHA-256, but 256-bit RSA isn't even large enough to hold SHA-256 hashes.

BoringSSL is starting to raise the minimum accepted RSA key, starting
with a floor of 512-bit, which is incompatible with this test. Per the
comments, this test wasn't set because 256-bit RSA was important, but
because RSA keygen isn't very fast. While true, it's not so slow that we
can't just test sizes matching reality, so bump this up to 2048.

There are other WebCrypto WPTs that test 512-bit RSA. I've left them
alone for now, though when we go to a minimum of 768-bit (for
background, 768-bit RSA has also already been factored, as of 2009),
we'll need to fix those too.

Bug: boringssl:607
Change-Id: I45df7a74961350448b7820aee26fc961b8f99706
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4541362
Reviewed-by: Matt Mueller <mattm@chromium.org>
Auto-Submit: David Benjamin <davidben@chromium.org>
Commit-Queue: Matt Mueller <mattm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1145470}