[iOS] Attempt to fix app life cycle crash after backgrounding
This CL is an attempt to fix the following crash:
https://crash.corp.google.com/browse?q=ReportId%3D%2798e363f6815061a3%27

The crash happens when a UIScene is foregrounded but the incognito
browser for it has not been initialized. When the app is foregrounded,
the metrics mediator attempts to count the active tabs in the incognito
browser of the scene to log metrics, but since the browser itself does
not exist, the code is essentially calling `nil->GetWebStateList()`.

Although the cause of this scenario is not yet identified, judging from
the breadcrumbs of crash reports with the same stack trace, this might
happen when the app is backgrounded (but not yet unattached) before
foregrounding again.

The `browserProviderInterface` for each scene state is implemented by
a `BrowserViewWrangler` object. The browser provider object and its
respective "browser" is created immediately after the browser view
wrangler is created, and the three objects are also deallocated at the
same time. Therefore, we can assume that when `activeScene.browser
ProviderInterface` exists, `activeScene.browserProviderInterface.
currentBrowserProvider.browser` would NOT be nil.

Bug: 1427703
Change-Id: I59a5567dce9550f29d9e3c95211ad083251f9220
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4538099
Commit-Queue: Ginny Huang <ginnyhuang@chromium.org>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1145492}